backend/requirements.txt
#requirements.txt
python-decouple
pyrebase4
firebase-admin
django-cors-headers
whitenoise
drf-yasg
pyjwt